fix examples dont build
